feat: add welcome email template
