+++ package.json
@@ -1,10 +1,10 @@
 {
   "name": "deepfocus",
   "version": "1.1.0",
-  "description": "An Electron application with Solid and TypeScript",
+  "description": "DeepFocus is a productivity tool that helps you stay focused and productive by tracking your time spent on your computer.",
   "main": "./out/main/index.js",
-  "author": "example.com",
-  "homepage": "https://electron-vite.org",
+  "author": "Timeo Williams",
+  "homepage": "https://deepFocus.cc",
   "type": "module",
   "scripts": {
     "format": "prettier --write .",
@@ -13,9 +13,9 @@
     "typecheck:web": "tsc --noEmit -p tsconfig.web.json --composite false",
     "typecheck": "npm run typecheck:node && npm run typecheck:web",
     "start": "electron-vite preview",
-    "dev": "electron-vite dev",
+    "dev": "electron-vite dev -- --inspect ",
     "build": "NODE_ENV=production npm run typecheck && electron-vite build",
-    "postinstall": "electron-builder install-app-deps",
+    "postinstall": "electron-builder install-app-deps && npx electron-rebuild",
     "build:unpack": "npm run build && electron-builder --dir",
     "build:win": "npm run build && electron-builder --win",
     "build:mac": "NODE_ENV=production electron-vite build && electron-builder --mac",
@@ -28,7 +28,7 @@
     "dotenv": "^16.4.5",
     "electron-store": "^10.0.0",
     "electron-updater": "^6.1.7",
-    "get-windows": "^9.1.1",
+    "get-windows": "^9.2.0",
     "node-schedule": "^2.1.1",
     "resend": "^3.4.0",
     "url": "^0.11.3"
